renamed covar to float location forecast
