+++ nodejs-server/models/floatLocationForecast.js
@@ -3,19 +3,6 @@ const Schema = mongoose.Schema;
 
 module.exports = {}
 
-const CovarSchema = Schema(
-  {
-    forcastDays: {type: Number, required: true },
-    _id: {type: String, required: true, max: 100},
-    features: [{type: Schema.Types.Mixed, required: true}],
-    geolocation: {type: Schema.Types.Mixed, required: true},
-    dLat: {type: Number, required: true },
-    dLong: {type :Number, required: true}
-  },
-);
-
-module.exports['covar'] = mongoose.model('covars', CovarSchema, 'covars');
-
 var geolocation = Schema({
   type: {
     type: String,
@@ -38,7 +25,7 @@ var sourceinfo = Schema({
   doi: {type: String, required: false}
 })
 
-const CovarianceSchema = Schema({
+const floatLocationForecastSchema = Schema({
   _id: {type: String, required: true},
   metadata: {type: [String], required: true},
   geolocation: {type: geolocation, required: true},
@@ -46,7 +33,7 @@ const CovarianceSchema = Schema({
   data: [{type: [Number]}]
 });
 
-const CovarianceMetaSchema = Schema({
+const floatLocationForecastMetaSchema = Schema({
   _id: {type: String, required: true},
   data_type: {type: String, required: true},
   date_updated_argovis: {type: Date, required: true},
@@ -56,5 +43,5 @@ const CovarianceMetaSchema = Schema({
   units: {type: [String], required: true},
 });
 
-module.exports['covariance'] = mongoose.model('covariance', CovarianceSchema, 'covariance');
-module.exports['covarianceMeta'] = mongoose.model('covarianceMeta', CovarianceMetaSchema , 'covarianceMeta');
+module.exports['floatLocationForecast'] = mongoose.model('covariance', floatLocationForecastSchema, 'covariance');
+module.exports['floatLocationForecastMeta'] = mongoose.model('covarianceMeta', floatLocationForecastMetaSchema , 'covarianceMeta');
